+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,8 +9,8 @@ EditorBuildSettings:
     path: Assets/Scenes/Title Screen.unity
     guid: 68f01b685d830cf46999a82834b7feeb
   - enabled: 1
-    path: Assets/Scenes/Controls.unity
-    guid: 9090bda3f41d8584eba32cbe4d698ac4
+    path: Assets/Scenes/Settings.unity
+    guid: 2f48bb89ee21e61468645d11aff9ad77
   - enabled: 1
     path: Assets/Scenes/Main scene.unity
     guid: 23d42cd4553f5b742b7810e89f06b771
